:recycle: [Refactor] noContents 이름 변경
